requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.61.0
python-docx==1.1.2
pdfplumber==0.11.4
